// src/empleado/aplicacion/dto/ConsultarOfertasLaborales.dto.ts
    export interface ConsultarOfertasLaboralesDTO {
        id: string
        titulo: string
        fecha: Date
        cargo: string
        sueldo: number
        duracionEstimadaValor: number
        duracionEstimadaEscala: string
        turnoTrabajo: string
        numeroVacantes: number
        nombreEmpresa: string
      }